+++ src/views/IncidentalsView/OngoingIncidentalsTable/CreateOngoingIncidentalsModal/CreateOngoingIncidentalsModal.tsx
@@ -3,15 +3,12 @@ import React from 'react';
 import { v4 as uuid } from 'uuid';
 import { ValidationConstraint } from '../../../../utils/validation/constraints';
 import Validator from '../../../../utils/validation/validator';
-import CurrencyInputField from '_/components/form/CurrencyInputField/CurrencyInputField';
 import DeductionTypeSelect from '_/components/form/DeductionTypeSelect/DeductionTypeSelect';
 import NumberTextField from '_/components/form/NumberTextField/NumberTextField';
 import GenericModal from '_/components/generic/GenericModal/GenericModal';
-import MonthYear from '_/extensions/date/month_year.extension';
 import useFormValidation from '_/hooks/useFormValidation/useFormValidation';
 import useIncidentalsState from '_/hooks/useIncidentalsState/useIncidentalsState';
 import { DeductionType } from '_/models/incidentals/deduction_type';
-import { CurrencyInCents } from '_/utils/currency/currency.utils';
 
 interface CreateOngoingIncidentalsModalProps {
   /**
@@ -35,11 +32,6 @@ interface CreateOngoingIncidentalsInput {
   name: string;
 
   /**
-   * Cost of the new incidentals
-   */
-  currentCost: CurrencyInCents;
-
-  /**
    * Deduction type of the new incidentals
    */
   deductionType: DeductionType;
@@ -53,7 +45,9 @@ interface CreateOngoingIncidentalsInput {
 /**
  * Modal that contains an input form to create new ongoing incidentals.
  */
-function CreateOngoingIncidentalsModal(props: CreateOngoingIncidentalsModalProps): JSX.Element {
+function CreateOngoingIncidentalsModal(
+  props: CreateOngoingIncidentalsModalProps,
+): JSX.Element {
   const { addOngoingIncidentals } = useIncidentalsState();
 
   const {
@@ -65,12 +59,10 @@ function CreateOngoingIncidentalsModal(props: CreateOngoingIncidentalsModalProps
   } = useFormValidation<CreateOngoingIncidentalsInput>({
     formValidator: new Validator<CreateOngoingIncidentalsInput>({
       name: ValidationConstraint.NoEmptyString,
-      currentCost: ValidationConstraint.Currency,
       invoiceInterval: ValidationConstraint.Defined,
     }),
     defaultFormInput: {
       name: '',
-      currentCost: undefined,
       deductionType: DeductionType.PerApartment,
       invoiceInterval: undefined,
     },
@@ -78,12 +70,7 @@ function CreateOngoingIncidentalsModal(props: CreateOngoingIncidentalsModalProps
       addOngoingIncidentals({
         id: uuid(),
         name: values.name,
-        costs: [
-          {
-            cost: values.currentCost,
-            date: new MonthYear().addMonths(-12),
-          },
-        ],
+        costs: [],
         deductionType: values.deductionType,
         invoiceInterval: values.invoiceInterval,
       });
@@ -103,7 +90,7 @@ function CreateOngoingIncidentalsModal(props: CreateOngoingIncidentalsModalProps
     >
       {/* Body */}
       <Grid container columnSpacing={2} rowSpacing={2}>
-        <Grid item xs={6}>
+        <Grid item xs={12}>
           <TextField
             fullWidth
             required
@@ -114,16 +101,6 @@ function CreateOngoingIncidentalsModal(props: CreateOngoingIncidentalsModalProps
             onChange={(event) => formInputSetters.name(event.target.value)}
             error={!!formErrors.name}
             helperText={formErrors.name}
-          />
-        </Grid>
-        <Grid item xs={6}>
-          <CurrencyInputField
-            required
-            id="cost"
-            label="Kosten"
-            value={formInput.currentCost}
-            onChange={formInputSetters.currentCost}
-            errorMessage={formErrors.currentCost}
           />
         </Grid>
         <Grid item xs={6}>